# api/openapi.yaml
openapi: 3.1.0
info:
  title: Touch API
  version: 1.0.0
  description: Standalone phone and watch touch delivery API with optional durable voice attachments.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/installations/{installationId}/push-tests:
    post:
      operationId: startPushTest
      parameters: [ { $ref: '#/components/parameters/InstallationId' }, { $ref: '#/components/parameters/InstallationHeader' } ]
      responses:
        '202': { description: Push challenge created, content: { application/json: { schema: { $ref: '#/components/schemas/PushTest' } } } }
        '401': { $ref: '#/components/responses/Error' }
...
  /v1/contacts:
    get:
      operationId: listContacts
      parameters: [ { $ref: '#/components/parameters/InstallationHeader' } ]
      responses:
        '200': { description: Saved and recent contacts, content: { application/json: { schema: { type: object, required: [contacts], properties: { contacts: { type: array, items: { $ref: '#/components/schemas/Contact' } } } } } } }
    post:
      operationId: addContact
      parameters: [ { $ref: '#/components/parameters/InstallationHeader' } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [username], properties: { username: { $ref: '#/components/schemas/Username' } } } } } }
      responses:
        '201': { description: Contact saved, content: { application/json: { schema: { $ref: '#/components/schemas/Contact' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /v1/contacts/{username}:
    delete:
      operationId: removeContact
      parameters: [ { name: username, in: path, required: true, schema: { $ref: '#/components/schemas/Username' } }, { $ref: '#/components/parameters/InstallationHeader' } ]
      responses: { '204': { description: Contact removed from saved contacts. } }
...
  /v1/admin/diagnostics:
    get:
      operationId: adminDiagnostics
      security: [ { adminBearer: [] } ]
      parameters:
        - { name: callId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: installationId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: messageId, in: query, schema: { $ref: '#/components/schemas/Uuid' } }
        - { name: severity, in: query, schema: { type: string } }
        - { name: sinceMs, in: query, schema: { type: integer, format: int64 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses: { '200': { description: Recent structured client diagnostics. }, '401': { $ref: '#/components/responses/Error' } }
...
  /v1/admin/calls:
    get:
      operationId: adminCalls
      security: [ { adminBearer: [] } ]
      responses: { '200': { description: Recent call sessions and transport counters. }, '401': { $ref: '#/components/responses/Error' } }
  /v1/admin/clients:
    get:
      operationId: adminClients
      security: [ { adminBearer: [] } ]
      responses: { '200': { description: Registered client summaries plus current ephemeral Live socket and stream state; no FCM tokens or media are returned. }, '401': { $ref: '#/components/responses/Error' } }
...
  /v1/admin/notifications:
    get:
      operationId: adminNotifications
      security: [ { adminBearer: [] } ]
      responses: { '200': { description: Recent FCM attempts and client delivery events. }, '401': { $ref: '#/components/responses/Error' } }
...
components:
  securitySchemes:
    adminBearer: { type: http, scheme: bearer }
  parameters:
    InstallationId:
      name: installationId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
    InstallationHeader:
      name: X-Installation-ID
      in: header
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
  responses:
    Error:
      description: Stable API error.
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }
  schemas:
    Uuid: { type: string, format: uuid }
    Username: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_.-]{2,31}$' }
...
    Contact:
      type: object
      required: [username, saved]
      properties:
        username: { $ref: '#/components/schemas/Username' }
        saved: { type: boolean }
        lastUsedAtMs: { type: integer, format: int64 }
    PushTest:
      type: object
      required: [testId, status, createdAtMs]
      properties:
        testId: { $ref: '#/components/schemas/Uuid' }
        status: { type: string, enum: [pending, delivered, timed_out, invalid_token, service_error] }
        createdAtMs: { type: integer, format: int64 }
        completedAtMs: { type: integer, format: int64 }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }